test(editor): split the Zed real-server guard
